+++ lib/nfg_ui/components/patterns/tile.rb
@@ -16,6 +16,14 @@ module NfgUi
           options.fetch(:heading, nil)
         end
 
+        def button
+          options.fetch(:button, nil)
+        end
+
+        def href
+          options.fetch(:href, nil)
+        end
+
         def component_family
           :tile
         end
@@ -28,7 +36,7 @@ module NfgUi
           super do
             if render_in_body
               if title.present?
-                concat(NfgUi::Components::Patterns::TileHeader.new({ title: title, icon: icon, collapsible: collapsible, collapse: ("#collapse_#{id}" if collapsible) }, view_context).render)
+                concat(NfgUi::Components::Patterns::TileHeader.new({ title: title, subtitle: subtitle, button: button, href: href, icon: icon, collapsible: collapsible, collapsed: collapsed, collapse: ("#collapse_#{id}" if collapsible) }, view_context).render)
               end
               if collapsible
                 concat(NfgUi::Components::Patterns::Collapse.new({ id: "collapse_#{id}", collapsed: collapsed }, view_context).render {
@@ -47,10 +55,14 @@ module NfgUi
           end
         end
 
+        def subtitle
+          options.fetch(:subtitle, nil)
+        end
+
         private
 
         def non_html_attribute_options
-          super.push(:heading, :render_in_body)
+          super.push(:heading, :render_in_body, :subtitle, :button, :href)
         end
       end
     end